requirements.txt
urllib3
websocket-client
pillow
python-dateutil
emoji
millify
pre-commit
colorama
flask
irc
pandas
pytz
locked-dict
plum-dispatch
